% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{create_label}
\alias{create_label}
\title{Create label list from dataframe}
\usage{
create_label(df)
}
\arguments{
\item{df}{}
}
\value{
A list of name-label pairs
}
\description{
Create label list from dataframe
}
\examples{
create_label(iris)
}
