% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_setup.R
\name{dir.setup}
\alias{dir.setup}
\title{Project directories setup}
\usage{
dir.setup()
}
\value{
N/A
}
\description{
Project directories setup
}
\seealso{
Other proj: 
\code{\link{templates}()}
}
\concept{proj}
