% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{get_rc_iem}
\alias{get_rc_iem}
\title{Get REDCAP instrument-event mapping}
\usage{
get_rc_iem(url = "https://redcap.ccf.org/redcap/api/", token)
}
\arguments{
\item{url}{URL}

\item{token}{API token}
}
\value{
A tibble
}
\description{
Get REDCAP instrument-event mapping
}
\seealso{
Other redcap: 
\code{\link{get_rc_metadata}()},
\code{\link{get_rc_records}()},
\code{\link{split_crf}()}
}
\concept{redcap}
