% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{get_rc_metadata}
\alias{get_rc_metadata}
\title{Get REDCAP metadata}
\usage{
get_rc_metadata(url = "https://redcap.ccf.org/redcap/api/", token)
}
\arguments{
\item{url}{URL}

\item{token}{API token}
}
\value{
A tibble
}
\description{
Get REDCAP metadata
}
\seealso{
Other redcap: 
\code{\link{get_rc_iem}()},
\code{\link{get_rc_records}()},
\code{\link{split_crf}()}
}
\concept{redcap}
