% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{get_rc_records}
\alias{get_rc_records}
\title{Get REDCAP records}
\usage{
get_rc_records(url = "https://redcap.ccf.org/redcap/api/", token)
}
\arguments{
\item{url}{URL}

\item{token}{API token}
}
\value{
A tibble
}
\description{
Get REDCAP records
}
\seealso{
Other redcap: 
\code{\link{get_rc_iem}()},
\code{\link{get_rc_metadata}()},
\code{\link{split_crf}()}
}
\concept{redcap}
