% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{odds}
\alias{odds}
\title{Calculate odds}
\usage{
odds(p)
}
\arguments{
\item{p}{Probability}
}
\value{
Numeric vector of length 1
}
\description{
Calculate odds
}
\seealso{
Other stats: 
\code{\link{paste_median}()}
}
\concept{stats}
