% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_utils.R
\name{p_print}
\alias{p_print}
\title{Pipe logger}
\usage{
p_print(data, msg = NULL, fun = dim)
}
\arguments{
\item{data}{A dataframe}

\item{msg}{Message to print}

\item{fun}{Function to apply to dataframe}
}
\value{
A dataframe
}
\description{
Pipe logger
}
\examples{
library(dplyr)
iris \%>\% p_print("Dimensions of iris dataframe")
}
