% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{paste_bin}
\alias{paste_bin}
\title{Summary statistics, n (\%), from binary variable}
\usage{
paste_bin(x)
}
\arguments{
\item{x:}{A vector with only two unique elements (0, 1)}
}
\value{
A character vector
}
\description{
Summary statistics, n (\%), from binary variable
}
\examples{
paste_bin()
}
