% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tbl12ft}
\alias{tbl12ft}
\title{Create table 1}
\usage{
tbl12ft(vars = NULL, strata, data = NULL, smd = T, ...)
}
\arguments{
\item{vars}{Character vector of variables to summarize}

\item{strata}{Grouping variable}

\item{data}{Dataframe}

\item{smd}{Report ASD or not.}

\item{...}{}
}
\value{
A flextable object
}
\description{
Create table 1
}
