% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_reporting.R
\name{tidy_p}
\alias{tidy_p}
\title{Tidy broom tables}
\usage{
tidy_p(tidy_df, terms, dig_est = 2, dig_p = 3, conf.level)
}
\arguments{
\item{tidy_df}{Tibble}

\item{est_round}{Number of digits to round to}
}
\value{
A tibble
}
\description{
Tidy broom tables
}
