% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_utils.R
\name{var_is_bin}
\alias{var_is_bin}
\title{Check if variable is binary}
\usage{
var_is_bin(x)
}
\arguments{
\item{A}{numeric vector}
}
\description{
Check if variable is binary
}
\examples{
var_is_bin(c(0,0,0,1,NA))
}
