% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{paste_median}
\alias{paste_median}
\title{Median IQR from numeric vector}
\usage{
paste_median(x)
}
\arguments{
\item{x}{Any numeric vector}
}
\value{
A character vector
}
\description{
Median IQR from numeric vector
}
\examples{
paste_median(iris$Sepal.Length)
}
