% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_utils.R
\name{tidy_p}
\alias{tidy_p}
\title{Tidy broom tables}
\usage{
tidy_p(tidy_df, est_round)
}
\arguments{
\item{tidy_df}{Tibble}

\item{est_round}{Number of digits to round to}
}
\value{
A tibble
}
\description{
Tidy broom tables
}
